Fix warnings.

+++ source/base/desktop/mouse_cursor.cc
@@ -50,7 +50,7 @@ MouseCursor& MouseCursor::operator=(MouseCursor&& other) noexcept
 
 int MouseCursor::stride() const
 {
-    return size_.width() * sizeof(uint32_t);
+    return size_.width() * static_cast<int>(sizeof(uint32_t));
 }
 
 bool MouseCursor::equals(const MouseCursor& other)
