+++ source/base/peer/client_authenticator.cc
@@ -435,7 +435,7 @@ void ClientAuthenticator::sendSessionResponse()
 
     response->set_os_name(SysInfo::operatingSystemName());
     response->set_computer_name(SysInfo::computerName());
-    response->set_cpu_cores(SysInfo::processorThreads());
+    response->set_cpu_cores(static_cast<uint32_t>(SysInfo::processorThreads()));
 
     LOG(LS_INFO) << "Sending: SessionResponse";
     sendMessage(*response);
